Fix warning

+++ src/Patch.cpp
@@ -52,11 +52,11 @@ namespace midikraft {
 		data_[sysExIndex] = value;
 	}
 
-	Synth::PatchData Patch::blankOut(std::vector<Range<int>> const &blankoutZones, Synth::PatchData const &inputData)
+	Synth::PatchData Patch::blankOut(std::vector<Range<size_t>> const &blankoutZones, Synth::PatchData const &inputData)
 	{
 		auto dataCopy = inputData;
 		for (auto blankOut : blankoutZones) {
-			for (int i = blankOut.getStart(); i < blankOut.getEnd(); i++) {
+			for (size_t i = blankOut.getStart(); i < blankOut.getEnd(); i++) {
 				dataCopy[i] = 0;
 			}
 		}
